remove processNodes.js
